Inherit version from environment in sub-programs
When a version was found using package.json set it to the
NODE_SHIM_$PROGRAM_VERSION environment variable. This env var overrides
the version resolution process so that any sub-programs spawned will use
the same version. It can also be used to force a certain version if you
don't want it automatically resolved.